+++ src/Extensibility/SharpDetect.Plugins/Deadlock/Descriptors/Arguments/CapturedArgumentDescriptor.cs
