+++ src/Extensibility/SharpDetect.Plugins/Deadlock/Descriptors/Arguments/CapturedValue.cs
