+++ src/Extensibility/SharpDetect.Plugins/Deadlock/Descriptors/Arguments/CapturedValueDescriptor.cs
