+++ src/Extensibility/SharpDetect.Plugins/Deadlock/Descriptors/MethodRewritingDescriptor.cs
