+++ src/Extensibility/SharpDetect.Plugins/HappensBeforeOrderingPluginBase.cs
@@ -5,14 +5,12 @@ using Microsoft.Extensions.DependencyInjection;
 using SharpDetect.Core.Events;
 using SharpDetect.Core.Events.Profiler;
 using SharpDetect.Core.Metadata;
-using SharpDetect.Core.Plugins.Descriptors;
+using SharpDetect.Core.Plugins;
 using SharpDetect.Core.Plugins.Models;
-using SharpDetect.Core.Plugins.PluginBases.MethodDescriptors;
 using SharpDetect.Core.Serialization;
-using System.Collections.Immutable;
 using System.Runtime.InteropServices;
 
-namespace SharpDetect.Core.Plugins.PluginBases.OrderedEvents;
+namespace SharpDetect.Plugins;
 
 public abstract class HappensBeforeOrderingPluginBase : PluginBase
 {
@@ -192,9 +190,6 @@ public abstract class HappensBeforeOrderingPluginBase : PluginBase
         _eventsDeliveryContext.BlockEventsDeliveryForThreadWaitingForObject(threadId, lockObj);
         return false;
     }
-
-    protected static ImmutableArray<MethodDescriptor> GetRequiredMethodDescriptors()
-        => MonitorMethodDescriptors.GetAllMethods().ToImmutableArray();
 
     protected override void Visit(RecordedEventMetadata metadata, ThreadCreateRecordedEvent args)
     {
